- Lots of improvements - Added insert and delete functionality - Added some labels to viewContorller to demonstrate number of views in scrollView and queue

+++ HorizontalTableView/Source/HorizontalTableViewCell.h
@@ -12,6 +12,6 @@
 @interface HorizontalTableViewCell : UIView
 
 @property (nonatomic, strong) NSString *identifier;
-@property (nonatomic, assign) NSInteger index;
+@property (nonatomic, assign) int index;
 
 @end
